LPS-134986 portal-search-spi: finishing moving to portal-search-spi
