+++ modules/apps/portal-search/portal-search-spi/src/main/java/com/liferay/portal/search/spi/model/permission/RoleSetContributorContext.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.spi.model.permission.context;
+package com.liferay.portal.search.spi.model.permission;
 
 import java.util.Set;
